add referrer, initial landing page, and legacy review reason to user review info card
