+++ packages/lesswrong/lib/collections/users/fragments.ts
@@ -266,6 +266,8 @@ registerFragment(`
     reviewedByUserId
     reviewedAt
     signUpReCaptchaRating
+    mapLocation
+    profileImageId
     
     needsReview
     sunshineNotes
