+++ packages/lesswrong/server/callbacks/sunshineCallbackUtils.tsx
@@ -1,44 +1,17 @@
-import Users from "../../lib/collections/users/collection";
-import { getCurrentContentCount } from '../../components/sunshineDashboard/ModeratorActions';
-import { Comments } from "../../lib/collections/comments";
-import { ModeratorActions } from "../../lib/collections/moderatorActions";
-import { createAdminContext, createMutator, updateMutator } from "../vulcan-lib";
-import { RECENTLY_DOWNVOTED_CONTENT_ALERT, LOW_AVERAGE_KARMA_COMMENT_ALERT, isActionActive, NEGATIVE_KARMA_USER_ALERT, LOW_AVERAGE_KARMA_POST_ALERT } from "../../lib/collections/moderatorActions/schema";
-import { forumTypeSetting } from "../../lib/instanceSettings";
-import { Posts } from "../../lib/collections/posts";
-import { isLowAverageKarmaContent } from "../../lib/collections/moderatorActions/helpers";
 import { CommentModeratorActions } from "../../lib/collections/commentModeratorActions/collection";
 import { DOWNVOTED_COMMENT_ALERT } from "../../lib/collections/commentModeratorActions/schema";
+import { Comments } from "../../lib/collections/comments";
+import { ModeratorActions } from "../../lib/collections/moderatorActions";
+import { getReasonForReview, isLowAverageKarmaContent } from "../../lib/collections/moderatorActions/helpers";
+import { isActionActive, LOW_AVERAGE_KARMA_COMMENT_ALERT, LOW_AVERAGE_KARMA_POST_ALERT, NEGATIVE_KARMA_USER_ALERT, RECENTLY_DOWNVOTED_CONTENT_ALERT } from "../../lib/collections/moderatorActions/schema";
+import { Posts } from "../../lib/collections/posts";
+import Users from "../../lib/collections/users/collection";
+import { createAdminContext, createMutator, updateMutator } from "../vulcan-lib";
 
-/** This function contains all logic for determining whether a given user needs review in the moderation sidebar.
- * 
+/** 
+ * This function contains all logic for determining whether a given user needs review in the moderation sidebar.
  * It's important this this only be be added to async callbacks on posts and comments, so that postCount and commentCount have time to update first
  */
-
-export function getReasonForReview(user: DbUser, override?: true) {
-  if (override && forumTypeSetting.get() === 'LessWrong') return 'override';
-
-  const fullyReviewed = user.reviewedByUserId && !user.snoozedUntilContentCount;
-  const neverReviewed = !user.reviewedByUserId;
-  const snoozed = user.reviewedByUserId && user.snoozedUntilContentCount;
-
-  if (fullyReviewed) return 'alreadyApproved';
-
-  if (neverReviewed) {
-    if (user.voteCount > 20) return 'voteCount';
-    if (user.mapLocation) return 'mapLocation';
-    if (user.postCount) return 'firstPost';
-    if (user.commentCount) return 'firstComment';
-    if (user.biography?.html) return 'bio';
-    if (user.profileImageId) return 'profileImage';  
-  } else if (snoozed) {
-    const contentCount = getCurrentContentCount(user);
-    if (contentCount >= user.snoozedUntilContentCount) return 'newContent';
-  }
-
-  return 'noReview';
-}
-
 export async function triggerReviewIfNeeded(userId: string, override?: true) {
   const user = await Users.findOne({ _id: userId });
   if (!user)
